# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bridge:
    get:
      operationId: getBridge
      tags: [bridge]
      summary: Get the consolidated EBITDA bridge
      description: Returns the ordered waterfall steps of the EBITDA bridge for a source/target scenario pair, including start/end totals and cumulative values.
      parameters:
        - $ref: "#/components/parameters/SourceScenario"
        - $ref: "#/components/parameters/TargetScenario"
      responses:
        "200":
          description: Bridge waterfall
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bridge"
        "400":
          description: Source and target periods have different granularity (year vs quarter vs month)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: No bridge for the selected scenario pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /bridge/components/{key}:
    get:
      operationId: getBridgeComponent
      tags: [bridge]
      summary: Get drill-down details for a bridge component
      description: Returns the supporting detail lines (from the Cálculo sheet) for one bridge component.
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/SourceScenario"
        - $ref: "#/components/parameters/TargetScenario"
      responses:
        "200":
          description: Component detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BridgeComponentDetail"
        "404":
          description: Component not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /bridge/tables:
    get:
      operationId: getBridgeTables
      tags: [bridge]
      summary: Get detailed tables per bridge driver
      description: Returns the detailed tables per driver (sales by product, fixed cost by category, input prices by item, usage, forex composition, stock/others), mirroring the Excel workbook driver tabs.
      parameters:
        - $ref: "#/components/parameters/SourceScenario"
        - $ref: "#/components/parameters/TargetScenario"
      responses:
        "200":
          description: Detailed tables
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BridgeTables"
        "400":
          description: Source and target periods have different granularity (year vs quarter vs month)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: No bridge for the selected scenario pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /bridge/simulate:
    post:
      operationId: simulateBridge
      tags: [bridge]
      summary: Simulate a what-if scenario from a natural-language prompt
      description: Interprets a Portuguese natural-language prompt (e.g. "se a venda de Calvert no MRF7 for 10% maior") into structured adjustments, applies them to the raw data of the selected scenario pair and recomputes the bridge.
      parameters:
        - $ref: "#/components/parameters/SourceScenario"
        - $ref: "#/components/parameters/TargetScenario"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateBridgeBody"
      responses:
        "200":
          description: Simulated bridge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateBridgeResponse"
        "400":
          description: Prompt could not be interpreted, no adjustment matched, or period-kind mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: No bridge for the selected scenario pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /bridge/summary:
    get:
      operationId: getBridgeSummary
      tags: [bridge]
      summary: Get executive summary of the bridge
      description: Returns headline figures — start, end, total variation, largest positive and negative impacts.
      parameters:
        - $ref: "#/components/parameters/SourceScenario"
        - $ref: "#/components/parameters/TargetScenario"
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BridgeSummary"
        "400":
          description: Source and target periods have different granularity (year vs quarter vs month)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: No bridge for the selected scenario pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
  parameters:
    SourceScenario:
      name: source
      in: query
      required: false
      description: Source scenario id (version + period of origin). Defaults to the first version with imported data.
      schema:
        type: string
    TargetScenario:
      name: target
      in: query
      required: false
      description: Target scenario id (version + period of destination). Defaults to the most recent version with imported data.
      schema:
        type: string
  schemas:
...
    ErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BridgeStep:
      type: object
      description: One bar of the waterfall chart, in MUSD.
      properties:
        key:
          type: string
          description: Stable identifier of the component (e.g. vol_mix, selling_price).
        label:
          type: string
          description: Display label in Portuguese.
        value:
          type: number
          description: Impact in MUSD (positive or negative). For totals, the absolute level.
        cumulative:
          type: number
          description: Cumulative EBITDA level after this step, in MUSD.
        kind:
          type: string
          enum: [total_start, delta, total_end]
        hasDetail:
          type: boolean
          description: Whether a drill-down detail exists for this component.
      required: [key, label, value, cumulative, kind, hasDetail]
    Bridge:
      type: object
      properties:
        title:
          type: string
        unit:
          type: string
          description: Unit of the values (MUSD).
        steps:
          type: array
          items:
            $ref: "#/components/schemas/BridgeStep"
      required: [title, unit, steps]
    BridgeDetailLine:
      type: object
      description: One supporting line of a bridge component drill-down.
      properties:
        id:
          type: number
        label:
          type: string
        group:
          type: ["string", "null"]
          description: Optional sub-grouping inside the component (e.g. Volume vs Mix, product family).
        value:
          type: number
          description: Impact in MUSD.
        sortOrder:
          type: number
      required: [id, label, group, value, sortOrder]
    BridgeComponentDetail:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: number
          description: Total impact of the component in MUSD.
        unit:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/BridgeDetailLine"
      required: [key, label, value, unit, lines]
    BridgeSummary:
      type: object
      properties:
        startLabel:
          type: string
        startValue:
          type: number
        endLabel:
          type: string
        endValue:
          type: number
        totalVariation:
          type: number
        largestPositive:
          $ref: "#/components/schemas/BridgeImpactHighlight"
        largestNegative:
          $ref: "#/components/schemas/BridgeImpactHighlight"
        positiveTotal:
          type: number
          description: Sum of all positive impacts, MUSD.
        negativeTotal:
          type: number
          description: Sum of all negative impacts, MUSD.
      required:
        - startLabel
        - startValue
        - endLabel
        - endValue
        - totalVariation
        - largestPositive
        - largestNegative
        - positiveTotal
        - negativeTotal
    BridgeTableColumn:
      type: object
      description: One numeric column of a detailed table; the label includes the unit.
      properties:
        key:
          type: string
        label:
          type: string
      required: [key, label]
    BridgeTableRow:
      type: object
      description: One row of a detailed table. Values align with the columns array; null means not applicable.
      properties:
        label:
          type: string
        kind:
          type: string
          enum: [row, subtotal, total]
        group:
          type: ["string", "null"]
          description: Row group (e.g. Blacks, Controllable). Subtotal rows with a group are the group header; plain rows with a group are its detail lines.
        values:
          type: array
          items:
            type: ["number", "null"]
      required: [label, kind, values]
    BridgeTable:
      type: object
      description: A detailed table for one bridge driver (mirrors an Excel driver tab).
      properties:
        key:
          type: string
        title:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/BridgeTableColumn"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/BridgeTableRow"
      required: [key, title, columns, rows]
    BridgeTables:
      type: object
      properties:
        title:
          type: string
        unit:
          type: string
        tables:
          type: array
          items:
            $ref: "#/components/schemas/BridgeTable"
      required: [title, unit, tables]
    SimulateBridgeBody:
      type: object
      properties:
        prompt:
          type: string
          description: Natural-language what-if instruction in Portuguese.
      required: [prompt]
    SimulatedBridgeTableRow:
      type: object
      description: One row of a simulated detailed table, with the original values and change flags aligned to the columns.
      properties:
        label:
          type: string
        kind:
          type: string
          enum: [row, subtotal, total]
        group:
          type: ["string", "null"]
          description: Row group (e.g. Blacks, Controllable). Subtotal rows with a group are the group header; plain rows with a group are its detail lines.
        values:
          type: array
          description: Simulated values, aligned with the columns; null means not applicable.
          items:
            type: ["number", "null"]
        baseValues:
          type: array
          description: Original (pre-simulation) values, aligned with the columns.
          items:
            type: ["number", "null"]
        changed:
          type: array
          description: Whether each cell was affected by the simulation (aligned with the columns).
          items:
            type: boolean
      required: [label, kind, values, baseValues, changed]
    SimulatedBridgeTable:
      type: object
      description: A detailed table recomputed with the simulated data, including original values per cell.
      properties:
        key:
          type: string
        title:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/BridgeTableColumn"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/SimulatedBridgeTableRow"
      required: [key, title, columns, rows]
    SimulateBridgeResponse:
      type: object
      properties:
        title:
          type: string
        unit:
          type: string
        interpretation:
          type: string
          description: Human-readable summary of how the prompt was interpreted.
        adjustments:
          type: array
          description: Descriptions of the adjustments actually applied.
          items:
            type: string
        steps:
          type: array
          description: Simulated waterfall steps (same shape as the base bridge).
          items:
            $ref: "#/components/schemas/BridgeStep"
        deltaEbitda:
          type: number
          description: Difference between the simulated and the original final EBITDA, in MUSD.
        tables:
          type: array
          description: Detailed tables recomputed with the simulated data, with original values and change flags per cell.
          items:
            $ref: "#/components/schemas/SimulatedBridgeTable"
      required: [title, unit, interpretation, adjustments, steps, deltaEbitda, tables]
    BridgeImpactHighlight:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: number
      required: [key, label, value]